Add action and trigger for syncing GitHub orgs
This approach aims to reduce complexity by using a triggered action
specifically for syncing github orgs when a repo is updated/inserted.

I've re-used the Octokit setup from the integration class, as it has
some complicated logic for determining which app installation to use
when making requests. In the future I would hope to install apps into
loops and skip this step, but for now it made sense to re-use the code
instead of duplicating it.

Resolves https://github.com/product-os/jellyfish/issues/8784

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>